Reorganize matching code.
