+++ pipeline/extract/matches.py
@@ -1,8 +1,9 @@
 import pandas as pd
 import prefect
 from prefect import task
+
+from pipeline.matching.core.release import from_release_tag
 from pipeline.schema.match import Match
-from pipeline.utils.release import from_release_tag
 
 
 @task
