+++ pipeline/matching/algorithms/fallback/__init__.py
@@ -1,0 +1,5 @@
+from pipeline.matching.algorithms.fallback.generator import get_fallback_matches
+from pipeline.matching.algorithms.fallback.optimizers import (
+    best_effort_minimize_repeat_matches,
+)
+from pipeline.matching.algorithms.fallback.tasks import compute_fallback_matches
