+++ pipeline/matching/core/constants.py
@@ -1,0 +1,1 @@
+N_MEMBERS_FOR_FINAL_MATCH = 3
