+++ pipeline/schema/types.py
@@ -1,0 +1,5 @@
+import datetime
+
+Community = str
+ReleaseTag = str
+ReleaseDate = datetime.datetime
