+++ requirements-dev.txt
@@ -1,4 +1,5 @@
 black==22.3.0
 isort==5.10.1
 pre-commit==2.18.1
+pycln==1.3.3
 pytest==6.2.4
